uns negocinhos bobos arrumados (ACHO Q VAI DAR BOA)

+++ app/src/main/java/oscar/go/com/oscarapp/filme/ListaFilmesActivity.java
@@ -27,7 +27,7 @@ import oscar.go.com.oscarapp.utilities.HttpHandler;
 public class ListaFilmesActivity extends Activity {
 
     private SessionManager session;
-    private int codU;
+    private long codU;
     private String userName;
     private ProgressDialog pDialog;
     private ListView list;
